Refactor CLI initialization and register tools from new module to improve compatibility with newer Python versions
